rapi3/requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.1.5
django-cors-headers==4.0.0
django-upgrade==1.12.0
djangorestframework==3.14.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.3
tokenize-rt==5.0.0
tzdata==2022.7
urllib3==2.0.2
